Fix missing signs for triangular BLAS2 functions
The PENCIL wrappers for triangular BLAS2 functions with reverse vector
access lacked a minus sign.  This was because 'arrayCases' contains
the cases for both the matrix and the vector.  Unfortunately, the case
'n' was used for both the matrix and the vector, with different
meanings.  We change the 'n' case (non-unit-triangular) for matrices
to 'e', such that each different case now uses a different letter.

Change-Id: I010a2c8dce4afe9de7c1597fc963bd0c9fe5e746